strlen on known array size may be optimized away
Seen with gcc-8.3.0:  The compiler may assume that the return value of
strlen will not be larger than the buffer size strlen is run on, because
that would imply reading beyond the buffer.  This breaks when the buffer
may not contain the trailing zero on purpose, optimizing away our
subsequent limitation on the length returned by strlen.