MsCompiler: preprocess assembler before assembling
Although the MASM (ml.exe or ml64.exe) is able to do some preprocessing,
the syntax is distinct from the C '#ifdef' style.  We need to assume
assembler source files are using the '#ifdef' style, so we need to
preprocess with the C or C++ compiler before we can run the assembler.

Also, support -xassembler and -xassembler-with-cpp compiler flag,
default to assembler-with-cpp for .S and .asm files.