parity: implemented option --help
This also led to an attempt to forward compiler or linker options
(-X/-Y) without specifying a source file, so that one can request the
toolchain's help by issuing 'g++ -X /?' or 'g++ -Y /?'.

The mechanism works, but after showing the respective help parity
reports an error and I do not know how to avoid this.

Change-Id: I1954d6a23501e936378d60bb59bb2c33db696a88